chore: refactor database and update readme
