--- server/app.js
@@ -1,12 +1,0 @@
-import http from 'http';
-import log from 'utils/logger';
-import app from './index';
-/**
- * Start Express server
- */
-const server = http.createServer(app);
-server.listen(app.get('port'), () => {
-    log.info('Server is running at port %s', app.get('port'));
-});
-
-module.exports = server;
